fix: add extend argument type
