+++ packages/icons/src/types.d.ts
@@ -1,0 +1,1 @@
+export type IconFiles = Record<string, () => Promise<string>>
